removed views related to metric profiles page to separate file
